Remove custom_config parameter
Remove the custom_config parameter on rsyslog::server

The case statement didn't work on Puppet 4.x because it only works if
undef == ''. In puppet 4 this causes compiles to fail.

puppetlabs-modules doesn't use this parameter anyway, so it's easier to
remove it entirely than to fix it. The newer version of saz-rsyslog uses
a different implementation anyway, so if we ever upgrade to it this
issue won't recur.